requirements.txt
gradio>=4.0
ollama
chromadb
pypdf
python-docx
